Refactor Cliente and Citta classes to enhance customer purchasing logic; adjust patience calculations and update currency type to float

+++ TheMerchant/Mercante.cs
@@ -1,4 +1,4 @@
 public static class Mercante
 {
-    public static int Monete { get; set; } = 1000;
+    public static float Monete { get; set; } = 1000;
 }
